Move CfPWeb into Web/Sources/WebCfP and Web/Public
