--- CfPWeb/Package.swift
@@ -1,25 +1,0 @@
-// swift-tools-version: 6.3
-
-import PackageDescription
-
-let package = Package(
-  name: "CfPWeb",
-  platforms: [.macOS(.v15)],
-  products: [
-    .executable(name: "CfPWeb", targets: ["CfPWeb"])
-  ],
-  dependencies: [
-    .package(url: "https://github.com/elementary-swift/elementary.git", from: "0.7.1"),
-  ],
-  targets: [
-    .executableTarget(
-      name: "CfPWeb",
-      dependencies: [
-        .product(name: "Elementary", package: "elementary"),
-      ],
-      swiftSettings: [
-        .swiftLanguageMode(.v6)
-      ]
-    )
-  ]
-)
