+++ Web/Sources/WebCfP/Components/AppNavigation.swift
