+++ Web/Sources/WebCfP/Components/Organizer/OrganizerWorkshopApplicationsContent.swift
