+++ Web/Sources/WebCfP/Components/Organizer/OrganizerWorkshopResultsContent.swift
