+++ Web/Sources/WebCfP/Components/Organizer/OrganizerWorkshopsContent.swift
