+++ Web/Sources/WebCfP/Support/AppConfiguration.swift
