+++ Web/Sources/WebCfP/Support/StaticSiteBuilder.swift
